refactor: Move Origin/NS running check to util mod
